gdb: xtensa: support larger than 256 byte stack offsets
Stack offset for spilled registers can be larger than 256 bytes, e.g.
on ADSP platforms with additional floating point and SIMD registers.

+++ include/zephyr/arch/xtensa/gdbstub.h
@@ -57,7 +57,7 @@ struct xtensa_register {
 	 * Offset of saved register in stack frame.
 	 * 0 if not saved in stack frame.
 	 */
-	int8_t		stack_offset;
+	int16_t		stack_offset;
 
 	/** Sequence number */
 	uint8_t		seqno;
